Fix the number type. Add logging.

+++ src/main/java/com/example/calculatorapp/model/Calculation.java
@@ -10,22 +10,22 @@ public class Calculation {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "firstNumber")
-    private Long firstNumber;
+    private Double firstNumber;
     @Column(name = "secondNumber")
-    private Long secondNumber;
+    private Double secondNumber;
     @Column(name = "result")
-    private Long result;
+    private Double result;
 
     private String action;
 
-    public Calculation(Long id, Long firstNumber, Long secondNumber, String action) {
+    public Calculation(Long id, Double firstNumber, Double secondNumber, String action) {
         this.id = id;
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
         this.action = action;
     }
 
-    public Calculation(Long firstNumber, Long secondNumber, String action) {
+    public Calculation(Double firstNumber, Double secondNumber, String action) {
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
         this.action = action;
@@ -34,18 +34,18 @@ public class Calculation {
     public Calculation() {
     }
 
-    public Calculation(Long firstNumber, Long secondNumber) {
+    public Calculation(Double firstNumber, Double secondNumber) {
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
     }
 
-    public Calculation(Long firstNumber, Long secondNumber, Long result) {
+    public Calculation(Double firstNumber, Double secondNumber, Double result) {
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
         this.result = result;
     }
 
-    public Calculation(Long id, Long firstNumber, Long secondNumber, Long result) {
+    public Calculation(Long id, Double firstNumber, Double secondNumber, Double result) {
         this.id = id;
         this.firstNumber = firstNumber;
         this.secondNumber = secondNumber;
@@ -60,27 +60,27 @@ public class Calculation {
         this.id = id;
     }
 
-    public Long getFirstNumber() {
+    public Double getFirstNumber() {
         return firstNumber;
     }
 
-    public void setFirstNumber(Long firstNumber) {
+    public void setFirstNumber(Double firstNumber) {
         this.firstNumber = firstNumber;
     }
 
-    public Long getSecondNumber() {
+    public Double getSecondNumber() {
         return secondNumber;
     }
 
-    public void setSecondNumber(Long secondNumber) {
+    public void setSecondNumber(Double secondNumber) {
         this.secondNumber = secondNumber;
     }
 
-    public Long getResult() {
+    public Double getResult() {
         return result;
     }
 
-    public void setResult(Long result) {
+    public void setResult(Double result) {
         this.result = result;
     }
 
@@ -92,22 +92,22 @@ public class Calculation {
         this.action = action;
     }
 
-    public Long plus() {
+    public double plus() {
         result = firstNumber + secondNumber;
         return result;
     }
 
-    public Long minus() {
+    public double minus() {
         result =  firstNumber - secondNumber;
         return result;
     }
 
-    public Long multiply() {
+    public double multiply() {
         result = firstNumber * secondNumber;
         return result;
     }
 
-    public Long division() {
+    public double division() {
         result = firstNumber / secondNumber;
         return result;
     }
